:twisted_rightwards_arrows: Move code from the github-actions repository here
It made sense for Cloud Platform to maintain this github action as Operations Engineering have no experience with the domain or the language. 0a18a518b4fd0e2e0ad981b1489c77bd90dbbd2e is the sha of the last commit in the github-actions repository, if you need to reference it. I did not personally write the code in this commit, but I did copy it over.